+++ tests/test_github_api.py
@@ -3,7 +3,7 @@ import typing as ty
 import pytest
 import requests
 
-from cs3560cli.github import GitHubApi
+from cs3560cli.github import GitHubApi, is_team_path
 
 
 class MockSuccessfulGetTeamIdResponse:
@@ -26,6 +26,11 @@ class MockSuccessfulInviteResponse:
         return {}
 
 
+def test_is_team_path() -> None:
+    assert is_team_path("A/B")
+    assert not is_team_path("A B")
+
+
 def test_githubapi_get_team_id_from_slug(monkeypatch: pytest.MonkeyPatch) -> None:
     def mock_get(*args, **kwargs):
         return MockSuccessfulGetTeamIdResponse()
@@ -34,6 +39,18 @@ def test_githubapi_get_team_id_from_slug(monkeypatch: pytest.MonkeyPatch) -> Non
 
     client = GitHubApi(token="fake-token")
     assert client.get_team_id_from_slug("OU-CS3560", "entire-class-20f") == 1000
+
+
+def test_githubapi_get_team_id_from_team_path(monkeypatch: pytest.MonkeyPatch) -> None:
+    def mock_get(*args, **kwargs):
+        return MockSuccessfulGetTeamIdResponse()
+
+    monkeypatch.setattr(requests, "get", mock_get)
+
+    client = GitHubApi(token="fake-token")
+    assert client.get_team_id_from_team_path("OU-CS3560/entire-class-20f") == 1000
+    with pytest.raises(ValueError):
+        client.get_team_id_from_team_path("OU-CS3560 entire-class-20f")
 
 
 def test_githubapi_invite_to_org(monkeypatch: pytest.MonkeyPatch) -> None: